pytorch/vlm-models/pipeline-code/requirements.txt
torch==2.5.0
transformers
datasets
accelerate
evaluate
bitsandbytes
trl
peft
protobuf
sentencepiece
argparse
pathlib
dotenv
pydantic-settings
wandb
pillow
torchvision
bert_score
nltk
absl-py
rouge_score
timm
pyyaml